feat: id の型を Int から Long に変更

+++ app/src/main/java/app/doggy/newmybrary/data/db/dao/AuthorDao.kt
@@ -6,8 +6,8 @@ import app.doggy.newmybrary.data.db.entity.AuthorEntity
 
 interface AuthorDao {
   @Insert
-  fun insert(author: AuthorEntity): Int
+  fun insert(author: AuthorEntity): Long
 
   @Update
-  fun update(author: AuthorEntity): Int
+  fun update(author: AuthorEntity): Long
 }
