+++ app/src/main/java/app/doggy/newmybrary/data/db/entity/BookAuthorCrossRef.kt
@@ -5,6 +5,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["book_id", "author_id"])
 data class BookAuthorCrossRef(
-  @ColumnInfo(name = "book_id") val bookId: Int,
-  @ColumnInfo(name = "author_id") val authorId: Int,
+  @ColumnInfo(name = "book_id") val bookId: Long,
+  @ColumnInfo(name = "author_id") val authorId: Long,
 )
